requirements.txt
Flask==3.0.2
Flask-SocketIO==5.3.6
gevent==24.2.1
gevent-websocket==0.10.1
greenlet==3.0.3
Jinja2==3.1.3
numpy==1.26.4
opencv-python==4.9.0.80
polling2==0.5.0
pyfcm==1.5.4
python-dotenv==1.0.1
python-multipart==0.0.9
python-socketio==5.11.1
requests==2.31.0
selenium==4.17.2
uvicorn==0.27.1
webdriver-manager==4.0.1
